Packaged alias redirect as a plugin
